generate_dot.py initializes EventManager
Also initializes StatsManager. This is needed because Flows and other
objects reference the global EventManager and StatsManager during
construction, and generate_dot.py initializes Flow and other objects.
