+++ host.py
@@ -7,25 +7,26 @@ class Host(Device):
     def __init__(self, id_):
         Device.__init__(self, id_)
         self.endpoint_for_router = None
-        if globals_.stats_manager:
-            self.bits_sent_graph_tag = globals_.stats_manager.new_graph(
-                    title='Total Bits Sent by %s' % self.id_,
-                    ylabel='Total Bits'
-            )
-            self.bit_rate_sent_graph_tag = globals_.stats_manager.new_graph(
-                    title='Rate of Data Sent by %s' % self.id_,
-                    ylabel='Rate (bits/sec)',
-                    is_rate=True
-            )
-            self.bits_received_graph_tag = globals_.stats_manager.new_graph(
-                    title='Total Bits Received by %s' % self.id_,
-                    ylabel='Total Bits'
-            )
-            self.bit_rate_received_graph_tag = globals_.stats_manager.new_graph(
-                    title='Rate of Data Received by %s' % self.id_,
-                    ylabel='Rate (bits/sec)',
-                    is_rate=True
-            )
+
+        self.bits_sent_graph_tag = globals_.stats_manager.new_graph(
+                title='Total Bits Sent by %s' % self.id_,
+                ylabel='Total Bits'
+        )
+        self.bit_rate_sent_graph_tag = globals_.stats_manager.new_graph(
+                title='Rate of Data Sent by %s' % self.id_,
+                ylabel='Rate (bits/sec)',
+                is_rate=True
+        )
+        self.bits_received_graph_tag = globals_.stats_manager.new_graph(
+                title='Total Bits Received by %s' % self.id_,
+                ylabel='Total Bits'
+        )
+        self.bit_rate_received_graph_tag = globals_.stats_manager.new_graph(
+                title='Rate of Data Received by %s' % self.id_,
+                ylabel='Rate (bits/sec)',
+                is_rate=True
+        )
+
         self.bits_sent = 0
         self.bits_received = 0
