+++ run.py
@@ -6,6 +6,7 @@ import globals_
 from event_manager import EventManager
 from stats_manager import StatsManager
 from file_parser import FileParser
+from main_setup import main_setup
 
 
 def simulate(network):
@@ -16,19 +17,6 @@ def simulate(network):
     globals_.event_manager.register_network(network)
     globals_.event_manager.run()
 
-
-def main_setup(output_name):
-    '''
-    This should be called at the beginning of main if you're running unit tests
-    or simulations.
-    It seeds the random number generator and sets the global variable
-    |globals_.event_manager|.
-    We need to set |globals_.event_manager| for most class definitions and helper
-    functions to make sense.
-    '''
-    random.seed(0)
-    globals_.event_manager = EventManager(output_name)
-    globals_.stats_manager = StatsManager(output_name)
 
 if __name__ == '__main__':
     if len(sys.argv) != 2 and len(sys.argv) != 3:
